Extract public common functions
- Extracted public common functions from ovsp4rt.cc and moved them
  to ovsp4rt_public.cc.
