+++ ovs-p4rt/sidecar/common/CMakeLists.txt
@@ -9,6 +9,7 @@ target_sources(ovs_sidecar_o PUBLIC
     ovsp4rt_credentials.cc
     ovsp4rt_credentials.h
     ovsp4rt_private.h
+    ovsp4rt_public.cc
     ovsp4rt_session.cc
     ovsp4rt_session.h
     ovsp4rt_utils.cc
